+++ restapis/controllers/document/documentDownload.js
@@ -1,6 +1,8 @@
 'use strict';
 const documentService = require('./documentMongoDB');
-const s3 = require('./documentS3');
+const {s3} = require('decompany-common-utils');
+const {region, s3Config} = require('decompany-app-properties');
+//const s3 = require('./documentS3');
 
 module.exports.handler = async (event, context, callback) => {
   console.log(event);
@@ -34,8 +36,9 @@ module.exports.handler = async (event, context, callback) => {
 
   const documentName = document.documentName;
   const ext  = documentName.substring(documentName.lastIndexOf(".") + 1, documentName.length).toLowerCase();
-  const signedUrl = s3.generateSignedUrl(document.accountId, document.documentId, ext);
-  
+  //const signedUrl = s3.generateSignedUrl(document.accountId, document.documentId, ext);
+  const documentKey = `FILE/${document.accountId}/${document.documentId}.${ext}`;
+  const signedUrl = await s3.signedDownloadUrl2({region: region, bucket: s3Config.document, key: documentKey, signedUrlExpireSeconds: 60});
   return JSON.stringify({
     success: true,
     downloadUrl: signedUrl,
