+++ restapis/test/documentRegist.js
@@ -1,13 +1,13 @@
 'use strict';
 
-// tests for trackingExport
+// tests for documentTracking
 // Generated by serverless-mocha-plugin
 
 const mochaPlugin = require('serverless-mocha-plugin');
 const expect = mochaPlugin.chai.expect;
-let wrapped = mochaPlugin.getWrapper('trackingExport', '/controllers/tracking/trackingExport.js', 'handler');
+let wrapped = mochaPlugin.getWrapper('documentRegist', '/controllers/document/documentRegist.js', 'handler');
 
-describe('trackingExport', () => {
+describe('documentRegist', () => {
   before((done) => {
     done();
   });
@@ -15,11 +15,19 @@ describe('trackingExport', () => {
   it('implement tests here', () => {
     const event = {
       principalId: "google-oauth2|101778494068951192848",
-      query: {
-        documentId: "5bab09df70ed4307aad667300fc0c95c"
+      body: {
+        filename: "dev_test.pdf",
+        size: "0",
+        tags: ["test", "dev"],
+        title: "dev_test",
+        desc: "dev test",
+        useTracking: true,
+        forceTracking: true,
+        isDownload: true
       }
     }
     return wrapped.run(event).then((response) => {
+      console.log("");
       console.log(response);
       expect(response).to.not.be.empty;
     });
