--- restapis/test/documentTracking.js
@@ -1,20 +1,0 @@
-'use strict';
-
-// tests for documentTracking
-// Generated by serverless-mocha-plugin
-
-const mochaPlugin = require('serverless-mocha-plugin');
-const expect = mochaPlugin.chai.expect;
-let wrapped = mochaPlugin.getWrapper('documentTracking', '/controllers/document/tracking.js', 'handler');
-
-describe('documentTracking', () => {
-  before((done) => {
-    done();
-  });
-
-  it('implement tests here', () => {
-    return wrapped.run({}).then((response) => {
-      expect(response).to.not.be.empty;
-    });
-  });
-});
